--- DynamicTable.xcodeproj/xcuserdata/ildarzalyalov.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "4AC2D4B3-9B74-4908-9E21-6FCF58B1A88D"
-   type = "1"
-   version = "2.0">
-</Bucket>
