--- DynamicTable/ViewController.swift
@@ -1,54 +1,0 @@
-//
-//  ViewController.swift
-//  DynamicTable
-//
-//  Created by Ильдар Залялов on 09/10/2019.
-//  Copyright © 2019 Ildar Zalyalov. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSource, UITextFieldDelegate {
-    
-    @IBOutlet weak var tableView: UITableView!
-    @IBOutlet weak var textField: UITextField!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        tableView.estimatedRowHeight = 100
-        textField.isSecureTextEntry = true
-        
-        textField.delegate = self
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 10
-    }
-       
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-       
-        let cell = tableView.dequeueReusableCell(withIdentifier: "cellIdentifier") as! CustomTableViewCell
-        
-        
-        if indexPath.row % 2 == 0 {
-             cell.configure(with: "Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.")
-        }
-        else {
-             cell.configure(with: "Hello world!")
-        }
-        
-        return cell 
-    }
-    
-    func textFieldDidBeginEditing(_ textField: UITextField) {
-        
-        
-    }
-    
-    func textFieldDidEndEditing(_ textField: UITextField) {
-       
-        
-    }
-}
-
