--- README.md
@@ -1,17 +1,0 @@
-# Lesson-4
-
-## Home work: 
-
-Переделать/сделать новое с ВК со следующими изменениями:   
-* Все на автолайаутах (отображение корректно на последних моделях + SE) 
-* Редактирование статуса пользователя через делегат на новом экране
-* Появляется верхняя часть вашей страницы, по нажатию на i переход на тот самый экран детального профиля.
-* Динамические записи на экране своей стены. 
-  - Используем одну ячейку для разных типов новостей. 
-* Добавить экран входа через логин пароль.   
-    - Логин и пароль валидируется: логин всегда email, все что не email пропускать не должно; пароль не меньше 6 символов, только латиница, содержит хотя бы 1 цифру.
-    - При ошибке валидации отображать UIAlertController с текстом ошибки. 
-* Для каждой пары логин пароль (читай юзера) своя инфа как на стене, так и о профиле (минимум 3 юзера)    
-
-### Итого: 
-У вас должно быть приложение которое начинается с экрана авторизации логина и пароля, при успешном вводе данных происходит переход на экран новостей, где каждую новость можно выбрать и посмотреть детально. Сверху вашей стены отображается не большая инфа о вас (как в ВК), где при нажатии на i можно перейти на экран детально информации профиля.
