+++ lesson4/lesson4.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:DynamicTable.xcodeproj">
+      location = "self:lesson4.xcodeproj">
    </FileRef>
 </Workspace>
