+++ lesson4/lesson4/Assets.xcassets/AppIcon.appiconset/Contents.json
