+++ lesson4/lesson4/Assets.xcassets/Contents.json
