+++ lesson4/lesson4/SceneDelegate.swift
@@ -1,9 +1,9 @@
 //
 //  SceneDelegate.swift
-//  DynamicTable
+//  lesson4
 //
-//  Created by Ильдар Залялов on 09/10/2019.
-//  Copyright © 2019 Ildar Zalyalov. All rights reserved.
+//  Created by Никита Лужбин on 11/10/2019.
+//  Copyright © 2019 Никита. All rights reserved.
 //
 
 import UIKit
@@ -46,6 +46,9 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Called as the scene transitions from the foreground to the background.
         // Use this method to save data, release shared resources, and store enough scene-specific state information
         // to restore the scene back to its current state.
+
+        // Save changes in the application's managed object context when the application transitions to the background.
+        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
     }
