+++ lesson4/lesson4/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  lesson4
+//
+//  Created by Никита Лужбин on 11/10/2019.
+//  Copyright © 2019 Никита. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
