+++ lesson4/lesson4/lesson4.xcdatamodeld/.xccurrentversion
@@ -2,13 +2,7 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>DynamicTable.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-	</dict>
+	<key>_XCCurrentVersionName</key>
+	<string>lesson4.xcdatamodel</string>
 </dict>
 </plist>
